Extract event-bus enums into models/events.py
models/common.py was a 869-line grab-bag (over the 800-line cap)
mixing event-bus enums, hardware pin enums, and catalog config-entry
models. Move EventType and StreamEvent — the event-bus / per-stream
protocol vocabularies — into a dedicated models/events.py, dropping
common.py to 490 lines.

Behaviour-free: events.py re-exports through models/__init__ so every
'from ...models import EventType' caller is unchanged; the lone direct
'from .common import EventType' in models/firmware.py is repointed.
