+++ esphome_device_builder/models/__init__.py
@@ -6,6 +6,7 @@ from .boards import *  # noqa: F403
 from .common import *  # noqa: F403
 from .components import *  # noqa: F403
 from .devices import *  # noqa: F403
+from .events import *  # noqa: F403
 from .firmware import *  # noqa: F403
 from .labels import *  # noqa: F403
 from .onboarding import *  # noqa: F403
